"Enable the setting of layer alpha value which used to blend with other layers"

+++ INI_Parser/ParseType.h
@@ -15,7 +15,7 @@ typedef struct
     unsigned int LayerPositionY;
     unsigned int LayerWidth;
     unsigned int LayerHeight;
-    unsigned int LayerAlpha;
+    float LayerAlpha;
     string LayerObjectModel;
     string LayerTexture;
 } LayerRenderType;
